feat: Refactored gRPC client implementation (#2)
This is a refactored copy of the Cerbos `client` package. The API has
remained unchanged except for slight changes to internal packages. For
most users the migration should be as simple as replacing `import
"github.com/cerbos/cerbos/client"` with `import
 client "github.com/cerbos/cerbos-sdk-go/cerbos"`.

Replaced `make` with `just` because it is much more ergonomic to use.

---------

Signed-off-by: Charith Ellawala <charith@cerbos.dev>